testing the location app/deploy-for-testing-face-main/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.19.5
opencv-python-headless==4.5.5.64
tensorflow==2.4.1
mtcnn==0.1.0
deepface==0.0.75
scikit-learn==1.0.2
imutils==0.5.4
python-dotenv==1.0.0
cloudinary==1.32.0
gunicorn==20.1.0
